[refactor] LoginViewModel 에서 애플 로그인 버튼 UI 리턴하는 로직 View 로 이동
