+++ BongBaek/BongBaek/Presentation/Login/ViewModel/LoginViewModel.swift
@@ -141,46 +141,21 @@ class LoginViewModel: ObservableObject {
 
 extension LoginViewModel {
     // MARK: - apple oauth
-    func requestAppleOauth() -> SignInWithAppleButton {
-        return SignInWithAppleButton(
-            onRequest: { request in
-                request.requestedScopes = [.fullName, .email]
-            },
-            onCompletion: { [weak self] result in
-                
-                guard let self = self else {
-                    print("self가 nil입니다.")
-                    return
-                }
-                
-                switch result {
-                case .success(let authResults):
-                    switch authResults.credential {
-                    case let appleIDCredential as ASAuthorizationAppleIDCredential:
-                        let identityToken = String(data: appleIDCredential.identityToken!, encoding: .utf8) ?? ""
-                        let authorizationCode = String(data: appleIDCredential.authorizationCode!, encoding: .utf8) ?? ""
-                        
+    func handleAppleLoginSuccess(identityToken: String, authorizationCode: String) {
 
-                        print("애플 인증 성공 - idToken: \(identityToken), authCode: \(authorizationCode)")
-                        
-                        Task {
-                            await self.authManager.loginWithApple(idToken: identityToken)
-                        }
-
-                
-                         
-                        self.authCode = authorizationCode
-                        self.idToken = identityToken
-                        
-                    default:
-                        break
-                    }
-                case .failure(let error):
-                    print(error.localizedDescription)
-                    print("애플 인증 에러 - \(error)")
-                }
-            }
-        )
+        print("애플 인증 성공 - idToken: \(identityToken), authCode: \(authorizationCode)")
+        
+        self.authCode = authorizationCode
+        self.idToken = identityToken
+        
+        Task {
+            await authManager.loginWithApple(idToken: identityToken)
+        }
+    }
+    
+    func handleAppleLoginFailure(error: Error) {
+        isLoading = false
+        errorMessage = error.localizedDescription
     }
 }
